+++ src/mod/codecs/mod_l16/mod_l16.c
@@ -52,10 +52,10 @@ static switch_status switch_raw_init(switch_codec *codec, switch_codec_flag flag
 static switch_status switch_raw_encode(switch_codec *codec,
 									   switch_codec *other_codec,
 									   void *decoded_data,
-									   size_t decoded_data_len,
-									   int decoded_rate,
+									   uint32_t decoded_data_len,
+									   uint32_t decoded_rate,
 									   void *encoded_data,
-									   size_t *encoded_data_len, int *encoded_rate, unsigned int *flag)
+									   uint32_t *encoded_data_len, uint32_t *encoded_rate, unsigned int *flag)
 {
 
 	/* NOOP indicates that the audio in is already the same as the audio out, so no conversion was necessary. */
@@ -70,10 +70,10 @@ static switch_status switch_raw_encode(switch_codec *codec,
 static switch_status switch_raw_decode(switch_codec *codec,
 									   switch_codec *other_codec,
 									   void *encoded_data,
-									   size_t encoded_data_len,
-									   int encoded_rate,
+									   uint32_t encoded_data_len,
+									   uint32_t encoded_rate,
 									   void *decoded_data,
-									   size_t *decoded_data_len, int *decoded_rate, unsigned int *flag)
+									   uint32_t *decoded_data_len, uint32_t *decoded_rate, unsigned int *flag)
 {
 	if (codec && other_codec && codec->implementation->samples_per_second != other_codec->implementation->samples_per_second) {
 		memcpy(decoded_data, encoded_data, encoded_data_len);
