+++ src/mod/codecs/mod_speex/mod_speex.c
@@ -160,10 +160,10 @@ static switch_status switch_speex_init(switch_codec *codec, switch_codec_flag fl
 static switch_status switch_speex_encode(switch_codec *codec,
 										 switch_codec *other_codec,
 										 void *decoded_data,
-										 size_t decoded_data_len,
-										 int decoded_rate,
+										 uint32_t decoded_data_len,
+										 uint32_t decoded_rate,
 										 void *encoded_data,
-										 size_t *encoded_data_len, int *encoded_rate, unsigned int *flag)
+										 uint32_t *encoded_data_len, uint32_t *encoded_rate, unsigned int *flag)
 {
 	struct speex_context *context = codec->private_info;
 	short *buf;
@@ -215,10 +215,10 @@ static switch_status switch_speex_encode(switch_codec *codec,
 static switch_status switch_speex_decode(switch_codec *codec,
 										 switch_codec *other_codec,
 										 void *encoded_data,
-										 size_t encoded_data_len,
-										 int encoded_rate,
+										 uint32_t encoded_data_len,
+										 uint32_t encoded_rate,
 										 void *decoded_data,
-										 size_t *decoded_data_len, int *decoded_rate, unsigned int *flag)
+										 uint32_t *decoded_data_len, uint32_t *decoded_rate, unsigned int *flag)
 {
 	struct speex_context *context = codec->private_info;
 	short *buf;
